refactor, add test case, google map working

+++ src/entity/entity.go
@@ -4,7 +4,7 @@ import "time"
 
 type Order struct {
 	ID          uint64    `gorm:"primary_key" json:"id"`
-	Distance    uint64    `gorm:"not null" json:"distance"`
+	Distance    int       `gorm:"not null" json:"distance"`
 	Status      string    `gorm:"type:varchar(10);not null" json:"status"`
 	OriginsLat  string    `json:"-"`
 	OriginsLong string    `json:"-"`
